requirements.txt
aiohttp==3.9.5
aiohttp-socks==0.8.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bottle==0.12.25
Brotli==1.1.0
browser-cookie3==0.19.1
cairocffi==1.7.0
CairoSVG==2.7.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
cssselect2==0.7.0
curl_cffi==0.6.4
defusedxml==0.7.1
deprecation==2.1.0
dnspython==2.6.1
duckduckgo_search==6.1.6
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
Flask==3.0.3
frozenlist==1.4.1
g4f==0.3.2.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
jeepney==0.8.0
Jinja2==3.1.4
loguru==0.7.2
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
nest-asyncio==1.6.0
orjson==3.10.4
packaging==24.1
pillow==10.3.0
platformdirs==4.2.2
plyer==2.1.0
proxy_tools==0.1.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
PyExecJS==1.5.1
Pygments==2.18.0
pyreqwest_impersonate==0.4.7
python-dotenv==1.0.1
python-multipart==0.0.9
python-socks==2.4.4
pywebview==5.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
tinycss2==1.3.0
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
xata==1.3.4
yarl==1.9.4
